pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "prevailingtealhashmap"
version = "0.1.0"
[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1"
"discord.py" = "^1.3"
[tool.poetry.dev-dependencies]
